printer: lambda bugfix.
* lib.c (obj_print_impl): Don't pass non-symbols to fboundp.
This causes a problem in the case where we are printing an
object like ((lambda ...) ...).  The car of this object is
the (lambda ...) form. When when pass this to fboundp, the
underlying function lookup mechanism wants to turn it into
a function object and tries to expand it. This can error out
if the lambda has bad syntax, which can happen because it's
just data that we are trying to print.