asm/vm/compiler: introduce jend pseudo-instruction.
The jend pseudo-instruction is a simple alias for end.  It
indicates a jumping end: an end that does not fall through to
the next instruction but could go somewhere else.

This is for "future correctness" as well as clarity.  The
difference is important in analysis of code into basic blocks.
Currently this won't make a difference because all the jend
instructions except for the one at the end of compiled
top-level form are followed by a label which kicks off a basic
block anyway.

* share/txr/stdlib/asm.tl (defopcode-alias): New macro.
(jend): New opcode, defined as alias for end.

* share/txr/stdlib/compiler.tl (comp-unwind-protect,
comp-lambda-impl, compile-toplevel): Use jend instruction for
a jumping end: the one after the protected code block of a
uwprot, the one at the end of a function, and the one at the
end of a top-level form.