txr: bugfix: give @(call) same semantics as direct call.
The @(call) directive is buggy in the following ways, which
cause an indirect call to behave differently from a direct call.

It creates a new context, and so if the opening of a data
source is deferred into the indirectly called function, that
data source is lost when the indirect call terminates.

Furthermore, if a data source is already established, there
is no progress through the data: two consecutive @(call ...)
directives operate on the same data.

It also fails to implement vertical to horizontal fallback; if
a function is not vertically defined, the directive fails.

* match.c (v_call): Rewrite the core logic in the following
way: we rewrite the indirect @(call) syntax into direct call
syntax, substitute that into c->spec, and then just
call v_fun.

* tests/008/call-2.expected: New file.

* tests/008/call-2.txr: New file. Test fails before this commit
because both calls are matching against the same "A" element
of the list.