txr: tighten keyword arg parsing in @(next).
This fixes the bug of not allowing @(next :list nil).
Also the bug of @(next :string) or @(next :string nil)
reporting a nonsensical error message, instead of correctly
complaining that nil isn't a string.

* match.c (v_next_impl): Capture the conses returned
by assoc, so we know whether a keyword is present.  Then use
those conses in tests which detect whether the keyword is
present, rather than the values, which could be nil.