compiler: move short end sequences before branch.
When a function ends like this:

           jmp label
           ...
   label:  end t42

The jmp can just be replaced by "end t42".

* share/txr/stdlib/optimize.tl (basic-blocks peephole-block):
Implement the pattern to replace a jump to a one or two
instruction sequence ending in a jend with that sequence.
Including the extra instruction helps if the target has not
been peepholed yet. In the pattern matching Ackermann case,
two instructions are promoted, one of which is eliminated.