compiler: merge duplicate jump blocks.
I've noticed that some duplicate short blocks are generated,
which look like this. Often they are consecutive, which
is why they are noticeable. These can become one block:

  mov t17 d3
  jmp label17
  mov t17 d3
  jmp label17
  mov t17 d3
  jmp label17

We identify identical blocks by looking for short instruction
sequences that end in an unconditional jmp, and then we group
duplicates in a hash table keyed on the instruction sequence.
We must ignore the label: the first instruction in each block
is a unique label.

We have to be careful about which ones to delete. Any block
that is entered from the top must be preserved. When these
blocks are identified, at least one block must remain that
is removable for the optimization to be able to do anything.

If all blocks are removable, we pick a leader which is
preserved. Otherwise we pick a leader from the preserved
blocks.  The non-preserved blocks are deleted, and all jumps
to them from other blocks are redirected to jumps to the
leader.

* share/txr/stdlib/compiler.tl (optimize): Call
merge-jump-thunks as the last pass.

* share/txr/stdlib/optimize.tl (basic-blocks
merge-jump-thunks): New method.