compiler: lift functional expressions to load-time.
The idea behind this optimization is that certain expressions
that only calculate functions can be hoisted to load time.

These expressions meet these criteria:

1. Are not already in a top-level or load-time context.

2. Are function calls to a standard library
   functional operator like chain andf, orf, juxt, ...

3. Do not access any variables.

3. Do not access any functions other than public (usr package)
   global functions in the standard library.

An example of such an expression might be:

   [chain cdr [iff symbolp list]]

If such an expression is embedded in a function, we don't want
the function to re-calculate it every time, which requires
time and generates garbage. We can transform it to the
equivalent of:

   (load-time [chain cdr [iff symbolp list]])

to have it calculated once.

* share/txr/stdlib/compiler.tl (%functional-funs%,
%functional%): New global variables.
(compiler comp-fun-form): After compiling the function call,
check for the conditions for lifting. If so, compile the form
again as a load-time literal. The logic is similar to how
lambdas are lifted to load-time, though the conditions are
different.