compiler: lift uslot and umethod forms too.
The uslot and umethod functions produce functions; and should
be lifted to load-time, if possible.

For instance, the .foo expression [mapcar .foo list]
translates to a (uslot 'foo) function call. This references no
variables, and so is liftable to load-time.

The umethod function is an extension of uslot that allows
partially applied method arguments to be carried. If those
arguments are all functional, the umethod call is liftable.

* share/txr/stdlib/compiler.tl (%functional-funs%): Include
umethod and uslot.