compiler: new late-peephole pass.
I noticed a lot of this in compiled code:

   (if reg label1)
   label2
   (jmp label3)
   label1

by inverting the test, this can be shortened to

   (ifq reg (t 0) label2)
   label1

We must keep label1 in case it is the target of other
branches. Also, we only perform this if label2 is unreachable:
the (jmp label3) instruction is reached only when the previous
if branch is not taken.

* share/txr/stdlib/optimize.tl (basic-blocks get-insns): Call
new late-peephole method.
(basic-blocks late-peephole): New method, incorporating the
above pattern.