lib: new functions join, join-with.
That old cat-str function is often a pain, requiring the
pieces as a list. We have a sys:fmt-join that is undocumented.
That functions is now exposed as usr:join, and documented.
Also introducing join-with that takes a separator as the
leftmost argument. Thus (op join-with "::") gives us
a function that joins pieces with :: in between.

* eval.c (eval_init): Regiser fmt_join function under join
symbol also. Register join-with.

* lib.c (join_with): New function.
(fmt_join): Now a wrapper for join_with that passes a nil
separator.

* lib.h (join_with): Declared.

* share/txr/stdlib/optimize.tl (basic-blocks join-blocks):
Rename the local function join, which now triggers a warning
about a standard function being redefined.

* txr.1: Redocumented cat-str, and documented join-with
and join.