compiler: improve end-propagating optimization.
When a jmp instruction is replaced by the end instruction
that it jumps to, or possibly by a two-instruction sequence
ending in end, there can be more opportunities to optimize.

For instance, the second reduction in a sequence like this:

     mov t3, t5        mov t3, t5
     jmp label   -->   end t3      -->    end t5
     ...
  label:
     end t3

* share/txr/stdlib/optimize.tl (basic-blocks peephole-block):
If the end-propagation is done, we change the linkage of the
current block to indicate that it has no next blocks. We
add it to the rescan list and set the recalc flag so the
liveness information is updated.