ffi: support float type as variadic argument.
The float type promotes to double when passed as
a variadic argument.

This patch adds internal FFI types which models that
promotion. It uses double for its C type, while still
performing the range checks for float.

Also, the types be-float and le-float are rejected from
being variadic arguments.

* share/txr/stdlib/ffi.tl (analyze-argtypes): Rewrite
function, adding validation and substitution for the
variadic part of the argument type list.  Map float type to
double, and reject be-float and le-float.

* txr.1: Documented.